pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
flask-sqlalchemy = "^3.1.1"
openai = "^1.51.0"
tenacity = "^9.0.0"
alembic = "^1.13.3"
flask-migrate = "^4.0.7"
requests = "^2.32.3"
pyperclip = "^1.9.0"
werkzeug = "^3.0.4"
python-dotenv = "^1.0.1"
quart = "^0.19.6"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"